wip: categories image

+++ src/Mymarket.Application/Features/Categories/Commands/Delete/DeleteCategoryCommand.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Mymarket.Application.Interfaces;
+using Mymarket.Application.Resources;
 
 namespace Mymarket.Application.Features.Categories.Commands.Delete;
 
@@ -9,7 +10,8 @@ public record DeleteCategoryCommand(
 ) : IRequest<Unit>;
 
 public class DeleteCategoryCommandHandler(
-    IApplicationDbContext context) : IRequestHandler<DeleteCategoryCommand, Unit>
+    IApplicationDbContext context,
+    IImageService imageService) : IRequestHandler<DeleteCategoryCommand, Unit>
 {
     public async Task<Unit> Handle(
         DeleteCategoryCommand request, CancellationToken cancellationToken)
@@ -17,8 +19,17 @@ public class DeleteCategoryCommandHandler(
         var category = await context.Categories
             .FirstOrDefaultAsync(c => c.Id == request.Id, cancellationToken);
 
+        if (category is null)
+            throw new KeyNotFoundException(SharedResources.IdDoesnotExist);
+
         context.Categories.Remove(category!);
         await context.SaveChangesAsync(cancellationToken);
+
+        if (category?.Logo is not null)
+        {
+            context.Images.Remove(category!.Logo);
+            await imageService.DeleteAsync(category!.Logo, cancellationToken);
+        }
 
         return Unit.Value;
     }
